fixing avx2 and non x86_64 issues
